requirements.txt
asgiref==3.6.0
Django==4.1.7
django-countries==7.5.1
django-multiselectfield==0.1.12
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
gunicorn==20.1.0